requirements.txt
pyyaml>=6.0
jsonschema>=4.0
pandas>=1.3.0
numpy>=1.20.0
python-dotenv>=0.19.0
pytest>=7.0.0
pytest-cov>=3.0.0
